requirements.txt

streamlit==1.25.0
pandas>=1.3.0
folium>=0.14.0
haversine>=2.8.0
Pillow>=8.0.0
